src/requirements.txt
pymodbus == 2.5.3
psycopg2-binary